+++ modulo2/labeddit/public/index.html
@@ -17,11 +17,11 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <style> html, body, #rrot{ margin: 0; padding: 0; width: 100%; min-height: 100vh;}
     </style>
-    <title>Labeddit</title>
+    <title>LabEddit</title>
     <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 
-<link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR:wght@300;700&family=Noto+Sans:wght@400;600&display=swap" rel="stylesheet">
+<link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR:wght@400;700&family=Noto+Sans:wght@400;600&display=swap" rel="stylesheet">
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
