+++ modulo2/labeddit/src/components/Header/Header.jsx
@@ -3,14 +3,21 @@ import {useState} from 'react'
 import { HeaderBar } from './styles';
 import { useNavigate } from 'react-router-dom';
 import {goToLogin} from '../../routes/coordinator'
+import x from '../../assets/x.png'
+import { useLocation } from 'react-router-dom';
 
 import Logo from '../../assets/logo.png'
 
 export default function Header() {
   const navigate = useNavigate()
   const token = localStorage.getItem('token')
-  const [rightButton, setButton] = useState(token ? 'Logout': 'Entrar')
- 
+  const [rightButton, setRightButton] = useState(token ? 'Logout': 'Entrar')
+
+const location = useLocation()
+const leftButton = () =>{
+  if (location.pathname == `/post/:id`)
+  return  <img src={x} />
+}
   const logout = () =>{
     localStorage.removeItem('token')
   }
@@ -18,7 +25,7 @@ export default function Header() {
   const rightButtonAction = () =>{
     if(token) {
       logout()
-      setButton('Entrar')
+      setRightButton('Entrar')
       goToLogin(navigate)
     }
     else {
@@ -26,7 +33,10 @@ export default function Header() {
     }
   }
   return (
+    
       <HeaderBar>
+    
+       {leftButton}
          <img src={Logo} alt="" />
         <button 
         onClick={rightButtonAction}
