+++ modulo2/labeddit/src/components/InputText/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const InputText = styled.input`
+width: 364px;
+height: 131px;
+background: #EDEDED;
+border-radius: 12px;
+border-style: none;
+`
