+++ modulo2/labeddit/src/constants/urls.js
@@ -1,4 +1,4 @@
 export const BASE_URL =
-    // 'https://labeddit.herokuapp.com'
-    'https://labeddit-2.herokuapp.com'
+    'https://labeddit.herokuapp.com'
+    //'https://labeddit-2.herokuapp.com'
