+++ modulo2/labeddit/src/pages/PostDetailPage/styles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const Xbutton = styled.img`
+position: absolute;
+top: 15px;
+left: 7%;
+
+:hover{
+        cursor: pointer;
+    }
+`
